[fix]- ajustadas colunas para decimal Id e Ref.

+++ AgendaWS/Domain/Repositories/IAgendaRepository.cs
@@ -8,9 +8,9 @@ namespace AgendaWS.Domain.Repositories
     {
         Task<Agenda> Salvar(Agenda agenda);
         Task<Agenda> Editar(Agenda agenda);
-        Task<bool> Excluir(int idAgenda);
+        Task<bool> Excluir(decimal idAgenda);
         Task<IEnumerable<Agenda>> Listar();
-        Task<Agenda> Obter(int idAgenda);
+        Task<Agenda> Obter(decimal idAgenda);
         Task<IEnumerable<Agenda>> ObterPorNome(string nome);
         Task<IEnumerable<Agenda>> ObterPorNumero(string numero);
     }
